Support devices that can't be reset from software
There are some devices out there that can't be reset from software, and need a
physical button pressed to put them into bootloader mode. On these devices,
trying to reset them is not going to work, so it is best skipped.

As such, this little change adds a `manualReset` flag to the board options,
which when set, allows `Avr109._reset` to return immediately, skipping the whole
reset dance. As a consequence, boards with the flag set, will have to be reset
by the user before calling the `_upload` (or `avrgirl.flash`) method.

Tested & works on the Shortcut keyboard, and fixes #111.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>